Fixed builds for UNIX based systems (Met the expectations of aggressive compiler flags)

+++ src/counting_sort.c
@@ -1,5 +1,6 @@
 #include "counting_sort.h"
 #include <assert.h>
+#include <limits.h>
 #include <stdio.h>
 
 int* cs_execute(const int* array, size_t length) {
@@ -20,7 +21,7 @@ int* cs_execute(const int* array, size_t length) {
         buffer[array[i]]++;
     }
 
-    for (size_t i = 1; i <= max; i++) {
+    for (int i = 1; i <= max; i++) {
         buffer[i] += buffer[i - 1];
     }
